pyproject.toml
[project]
name = "bloging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.48",
]
